bpf, lib: remove xdp specific header
Get rid of xdp.h header since lib/ should not contain anything
context type specific. Move parts we can reuse into lib/common.h.
